+++ spec/rubocop/cop/style/if_unless_modifier_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe RuboCop::Cop::Style::IfUnlessModifier do
   subject(:cop) { described_class.new(config) }
 
   let(:config) do
-    RuboCop::Config.new('Metrics/LineLength' => line_length_config)
+    RuboCop::Config.new('Layout/LineLength' => line_length_config)
   end
   let(:line_length_config) do
     {
@@ -27,7 +27,7 @@ RSpec.describe RuboCop::Cop::Style::IfUnlessModifier do
       'https://some.example.com/with/a/rather?long&and=very&complicated=path'
     end
 
-    context 'when Metrics/LineLength is enabled' do
+    context 'when Layout/LineLength is enabled' do
       it 'corrects it to normal form' do
         expect(source.length).to be(79) # That's 81 including indentation.
         expect_offense(<<~RUBY)
@@ -106,7 +106,7 @@ RSpec.describe RuboCop::Cop::Style::IfUnlessModifier do
       end
     end
 
-    context 'when Metrics/LineLength is disabled in configuration' do
+    context 'when Layout/LineLength is disabled in configuration' do
       let(:line_length_config) { { 'Enabled' => false, 'Max' => 80 } }
 
       it 'accepts' do
@@ -118,24 +118,24 @@ RSpec.describe RuboCop::Cop::Style::IfUnlessModifier do
       end
     end
 
-    context 'when Metrics/LineLength is disabled with enable/disable ' \
+    context 'when Layout/LineLength is disabled with enable/disable ' \
             'comments' do
       it 'accepts' do
         expect_no_offenses(<<~RUBY)
           def f
-            # rubocop:disable Metrics/LineLength
+            # rubocop:disable Layout/LineLength
             #{source}
-            # rubocop:enable Metrics/LineLength
+            # rubocop:enable Layout/LineLength
           end
         RUBY
       end
     end
 
-    context 'when Metrics/LineLength is disabled with an EOL comment' do
+    context 'when Layout/LineLength is disabled with an EOL comment' do
       it 'accepts' do
         expect_no_offenses(<<~RUBY)
           def f
-            #{source} # rubocop:disable Metrics/LineLength
+            #{source} # rubocop:disable Layout/LineLength
           end
         RUBY
       end
@@ -552,7 +552,7 @@ RSpec.describe RuboCop::Cop::Style::IfUnlessModifier do
             'Enabled' => false,
             'IndentationWidth' => 2
           },
-          'Metrics/LineLength' => { 'Max' => 10 + 12 } # 12 is indentation
+          'Layout/LineLength' => { 'Max' => 10 + 12 } # 12 is indentation
         )
       end
 
@@ -568,7 +568,7 @@ RSpec.describe RuboCop::Cop::Style::IfUnlessModifier do
           'Layout/Tab' => {
             'Enabled' => false
           },
-          'Metrics/LineLength' => { 'Max' => 10 + 6 } # 6 is indentation
+          'Layout/LineLength' => { 'Max' => 10 + 6 } # 6 is indentation
         )
       end
 
@@ -581,7 +581,7 @@ RSpec.describe RuboCop::Cop::Style::IfUnlessModifier do
           'Layout/Tab' => {
             'Enabled' => false
           },
-          'Metrics/LineLength' => { 'Max' => 10 + 12 } # 12 is indentation
+          'Layout/LineLength' => { 'Max' => 10 + 12 } # 12 is indentation
         )
       end
 
@@ -589,10 +589,10 @@ RSpec.describe RuboCop::Cop::Style::IfUnlessModifier do
     end
   end
 
-  context 'when Metrics/LineLength is disabled' do
+  context 'when Layout/LineLength is disabled' do
     let(:config) do
       RuboCop::Config.new(
-        'Metrics/LineLength' => {
+        'Layout/LineLength' => {
           'Enabled' => false,
           'Max' => 80
         }